requirements.txt
depthai
numpy
pymavlink
opencv-python
pyyaml
